// src/app/api/models/cuenta-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CuentaResponse {
  alumno_id?: number;
  cronograma_cuotas_id?: number;
  cuota?: string;
  estado?: 'PENDIENTE' | 'PAGADO' | 'ANULADO' | 'VENCIDO';
  fechaVencimiento?: string;
  monto?: number;
  periodoAcademico?: string;
  saldo?: number;
}